requirements.txt
streamlit
gspread
openai
google-cloud-texttospeech
google-generativeai
pydub
matplotlib
seaborn
pandas
oauth2client
